+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 * Partitioning supports original indices
 * Acero engine declaration
+* `Duration` format improvements
 
 ## [1.6](https://pypi.org/project/graphique/1.6/) - 2024-04-30
 ### Changed